requirements.txt
notion-client==2.2.1
anthropic==0.49.0
slack-sdk==3.34.0
python-dotenv==1.1.0
pytest==8.3.5
pytest-cov==6.1.0
